requirements.txt
pyyaml
regex
pandas
openpyxl
rich
